feat(): convert helpers module into command mixin
to extend common command methods
