+++ lib/paperclip_attributes/engine.rb
@@ -6,7 +6,7 @@ module PaperclipAttributes
 
     initializer "initialize" do
       require_relative "./error"
-      require_relative "./helpers"
+      require_relative "./command"
       require_relative "./commands/add_photo_dimensions"
       require_relative "./commands/add_photo_dominant_color"
       require_relative "./attributes"
